WSTEAM1-309: Refactor
